requirements.txt
six
pandas
numpy
matplotlib
requests
Pillow
requests_oauthlib